requirements.txt
pre-commit==3.0.4
black==23.1.0
isort==5.12.0
flake8==6.0.0
pandas==1.5.3
pysnowflake==0.1.3
snowflake-connector-python==3.0.4
snowflake-sqlalchemy==1.4.7
apache-airflow-providers-snowflake==4.1.0